move common.InterruptibleSleep to util
