+++ common/util/util.go
@@ -27,6 +27,7 @@
 package util
 
 import (
+	"context"
 	"sort"
 	"time"
 
@@ -171,4 +172,14 @@ func RepeatSlice[T any](xs []T, n int) []T {
 // Ptr returns a pointer to a copy of v.
 func Ptr[T any](v T) *T {
 	return &v
+}
+
+// InterruptibleSleep is like time.Sleep but can be interrupted by a context.
+func InterruptibleSleep(ctx context.Context, timeout time.Duration) {
+	timer := time.NewTimer(timeout)
+	defer timer.Stop()
+	select {
+	case <-timer.C:
+	case <-ctx.Done():
+	}
 }
